# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080/","description":"Generated server url"}],"paths":{"/guest":{"get":{"tags":["guest-controller"],"summary":"Get a guest by its id","operationId":"getGuest","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Found the Guest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuestEntity"}}}},"400":{"description":"Invalid id supplied"},"404":{"description":"Guest not found"}}},"put":{"tags":["guest-controller"],"summary":"Update a guest by Its id","operationId":"updateGuest","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"fName","in":"query","required":true,"schema":{"type":"string"}},{"name":"lName","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Update the Guest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuestEntity"}}}},"404":{"description":"Guest is not updated"},"400":{"description":"Invalid  Guest id supplied"}}},"post":{"tags":["guest-controller"],"summary":"Save a guest by GuestDTO","operationId":"saveGuest","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuestDTO"}}},"required":true},"responses":{"404":{"description":"Guest not saved"},"200":{"description":"Save the Guest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuestEntity"}}}},"400":{"description":"Invalid details supplied"}}},"delete":{"tags":["guest-controller"],"summary":"Delete a guest by its id","operationId":"deleteGuest","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Delete the Guest Sucessfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuestEntity"}}}},"400":{"description":"Invalid id supplied"},"404":{"description":"Guest not found"}}}},"/user":{"get":{"tags":["user-controller"],"operationId":"getUser","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}},"post":{"tags":["user-controller"],"operationId":"saveUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/train":{"get":{"tags":["train-controller"],"summary":"Get a train Record","operationId":"getTrain","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TrainDTO"}}}}}},"post":{"tags":["train-controller"],"summary":"Save a train Record","operationId":"saveTrain","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TrainEntity"}}}}}},"delete":{"tags":["train-controller"],"summary":"Delete a train Record","operationId":"deleteTrain","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/reservation":{"get":{"tags":["reservation-controller"],"operationId":"getReservation","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationDTO"}}}}}},"post":{"tags":["reservation-controller"],"operationId":"saveReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationEntity"}}}}}},"delete":{"tags":["reservation-controller"],"operationId":"deleteReservation","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/guestreservation":{"get":{"tags":["guest-reservation-controller"],"operationId":"getGuestReservations","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GuestReservationDTO"}}}}}},"post":{"tags":["guest-reservation-controller"],"operationId":"saveGuestReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuestReservationDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GuestReservationEntity"}}}}}}},"/users":{"get":{"tags":["user-controller"],"operationId":"readUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}}}}},"/trains":{"get":{"tags":["train-controller"],"summary":"Get All Train Records","operationId":"findAllTrain","parameters":[{"name":"start","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelTrainDTO"}}}}}}},"/reservations":{"get":{"tags":["reservation-controller"],"operationId":"readReservation","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReservationDTO"}}}}}}}},"/guests":{"get":{"tags":["guest-controller"],"summary":"Get a guest by its id","operationId":"findallguest","parameters":[{"name":"start","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"pageSize","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Found the Guest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuestEntity"}}}},"400":{"description":"Invalid id supplied"},"404":{"description":"Guest not found"}}}},"/guestreservations":{"get":{"tags":["guest-reservation-controller"],"operationId":"readGuestReservation","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GuestReservationDTO"}}}}}}}},"/getallreservationbyguestid":{"get":{"tags":["guest-controller"],"summary":"Get all reservation of a  guest by its id","operationId":"readguests","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"start","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"pageSize","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Found the Guest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuestEntity"}}}},"400":{"description":"Invalid id supplied"},"404":{"description":"Guest not found"}}}},"/guestReservation":{"delete":{"tags":["guest-reservation-controller"],"operationId":"deleteGuestReservation","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"GuestEntity":{"type":"object","properties":{"addDate":{"type":"string","format":"date-time"},"addedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"guestId":{"type":"integer","format":"int32"},"firstName":{"type":"string"},"lastName":{"type":"string"},"gender":{"type":"string"},"email":{"type":"string"},"deleted":{"type":"boolean"},"lastModifiedBy":{"type":"string"}}},"UserDTO":{"type":"object","properties":{"addDate":{"type":"string","format":"date-time"},"addedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"userId":{"type":"integer","format":"int32"},"firstName":{"type":"string"},"lastName":{"type":"string"},"gender":{"type":"string"},"email":{"type":"string"},"userpassword":{"type":"string"},"deleted":{"type":"boolean"},"lastModifiedBy":{"type":"string"}}},"UserEntity":{"type":"object","properties":{"addDate":{"type":"string","format":"date-time"},"addedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"userId":{"type":"integer","format":"int32"},"firstName":{"type":"string"},"lastName":{"type":"string"},"gender":{"type":"string"},"email":{"type":"string"},"userpassword":{"type":"string"},"deleted":{"type":"boolean"},"lastModifiedBy":{"type":"string"}}},"TrainDTO":{"type":"object","properties":{"addDate":{"type":"string","format":"date-time"},"addedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"trainDetailId":{"type":"integer","format":"int32"},"trainNo":{"type":"integer","format":"int32"},"startingStation":{"type":"string"},"lastStation":{"type":"string"},"departureDate":{"type":"string","format":"date-time"},"arrivalDate":{"type":"string","format":"date-time"},"deleted":{"type":"boolean"},"lastModifiedBy":{"type":"string"}}},"TrainEntity":{"type":"object","properties":{"addDate":{"type":"string","format":"date-time"},"addedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"trainDetailId":{"type":"integer","format":"int32"},"trainNo":{"type":"integer","format":"int32"},"startingStation":{"type":"string"},"lastStation":{"type":"string"},"departureDate":{"type":"string","format":"date-time"},"arrivalDate":{"type":"string","format":"date-time"},"deleted":{"type":"boolean"},"lastModifiedBy":{"type":"string"}}},"ReservationDTO":{"type":"object","properties":{"addDate":{"type":"string","format":"date-time"},"addedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"reservationId":{"type":"integer","format":"int32"},"pnrNo":{"type":"integer","format":"int64"},"trainNo":{"type":"integer","format":"int32"},"startingStation":{"type":"string"},"lastStation":{"type":"string"},"departureDate":{"type":"string","format":"date-time"},"arrivalDate":{"type":"string","format":"date-time"},"currentStatus":{"type":"boolean"},"deleted":{"type":"boolean"},"lastModifiedBy":{"type":"string"}}},"ReservationEntity":{"type":"object","properties":{"addDate":{"type":"string","format":"date-time"},"addedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"reservationId":{"type":"integer","format":"int32"},"pnrNo":{"type":"integer","format":"int64"},"trainNo":{"type":"integer","format":"int32"},"startingStation":{"type":"string"},"lastStation":{"type":"string"},"departureDate":{"type":"string","format":"date-time"},"arrivalDate":{"type":"string","format":"date-time"},"currentStatus":{"type":"boolean"},"deleted":{"type":"boolean"},"lastModifiedBy":{"type":"string"}}},"GuestReservationDTO":{"type":"object","properties":{"addDate":{"type":"string","format":"date-time"},"addedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"guestReservationId":{"type":"integer","format":"int32"},"bookedBy":{"type":"string"},"reservationId":{"type":"integer","format":"int32"},"guestId":{"type":"integer","format":"int32"},"currentStatus":{"type":"boolean"},"deleted":{"type":"boolean"},"lastModifiedBy":{"type":"string"}}},"GuestReservationEntity":{"type":"object","properties":{"addDate":{"type":"string","format":"date-time"},"addedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"guestReservationId":{"type":"integer","format":"int32"},"bookedBy":{"type":"string"},"reservationId":{"type":"integer","format":"int32"},"guestId":{"type":"integer","format":"int32"},"currentStatus":{"type":"boolean"},"deleted":{"type":"boolean"},"lastModifiedBy":{"type":"string"}}},"GuestDTO":{"type":"object","properties":{"addDate":{"type":"string","format":"date-time"},"addedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"firstName":{"type":"string"},"lastName":{"type":"string"},"gender":{"type":"string"},"email":{"type":"string"},"reservationList":{"type":"array","items":{"$ref":"#/components/schemas/ReservationDTO"}},"guestId":{"type":"integer","format":"int32"},"deleted":{"type":"boolean"},"lastModifiedBy":{"type":"string"}}},"EntityModelTrainDTO":{"type":"object","properties":{"addDate":{"type":"string","format":"date-time"},"addedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"trainDetailId":{"type":"integer","format":"int32"},"trainNo":{"type":"integer","format":"int32"},"startingStation":{"type":"string"},"lastStation":{"type":"string"},"departureDate":{"type":"string","format":"date-time"},"arrivalDate":{"type":"string","format":"date-time"},"deleted":{"type":"boolean"},"lastModifiedBy":{"type":"string"},"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}}}},"Link":{"type":"object","properties":{"rel":{"type":"string"},"href":{"type":"string"},"hreflang":{"type":"string"},"media":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"deprecation":{"type":"string"},"profile":{"type":"string"},"name":{"type":"string"}}},"PageMetadata":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int64"}}},"PagedModelEntityModelTrainDTO":{"type":"object","properties":{"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}},"content":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelTrainDTO"}},"page":{"$ref":"#/components/schemas/PageMetadata"}}}}}}