# src/main/java/com/rishi/userbooking/DTO/GuestDTO.java
package com.rishi.userbooking.DTO;

import java.util.List;

public class GuestDTO extends BaseDTO {
	/**
	 * 
	 */
	private static final long serialVersionUID = 478891755248735420L;
	private Integer GuestId;
	private String firstName;
	private String lastName;
	private String gender;
	private String email;
	private boolean isDeleted;
	private List<ReservationDTO> reservationList;
	
	public List<ReservationDTO> getReservationList() {
		return reservationList;
	}
	public void setReservationList(List<ReservationDTO> reservationList) {
		this.reservationList = reservationList;
	}
	public Integer  getGuestId() {
		return GuestId;
	}
	public void setGuestId(Integer guestId) {
		GuestId = guestId;
	}
	public String getFirstName() {
		return firstName;
	}
	public void setFirstName(String firstName) {
		this.firstName = firstName;
	}
	public String getLastName() {
		return lastName;
	}
	public void setLastName(String lastName) {
		this.lastName = lastName;
	}
	public String getGender() {
		return gender;
	}
	public void setGender(String gender) {
		this.gender = gender;
	}
	public String getEmail() {
		return email;
	}
	public void setEmail(String email) {
		this.email = email;
	}
	public boolean isDeleted() {
		return isDeleted;
	}
	public void setDeleted(boolean isDeleted) {
		this.isDeleted = isDeleted;
	}
	
	
	
}